refactor: _header-client 파일명에서 언더스코어 제거 (#142)
